[Feat] : 홈화면 response 수정 (#88)
