+++ src/main/java/com/potato/balbambalbam/data/repository/CardBookmarkRepository.java
@@ -3,6 +3,8 @@ package com.potato.balbambalbam.data.repository;
 import com.potato.balbambalbam.data.entity.CardBookmark;
 import com.potato.balbambalbam.data.entity.CardBookmarkId;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -18,4 +20,7 @@ public interface CardBookmarkRepository extends JpaRepository<CardBookmark, Card
     boolean existsByUserId(Long userId);
 
     List<CardBookmark> findAllByUserId(Long userId);
+
+    @Query("SELECT COUNT(cb) FROM card_bookmark cb WHERE cb.userId = :userId")
+    Long countByUserId(@Param("userId") Long userId);
 }
