+++ src/main/java/com/potato/balbambalbam/data/repository/CardScoreRepository.java
@@ -2,6 +2,8 @@ package com.potato.balbambalbam.data.repository;
 
 import com.potato.balbambalbam.data.entity.CardScore;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -16,4 +18,7 @@ public interface CardScoreRepository extends JpaRepository<CardScore, Long> {
     void deleteByUserId(Long userId);
 
     boolean existsByUserId(Long userId);
+
+    @Query("SELECT COUNT(cs) FROM card_score cs WHERE cs.userId = :userId AND cs.highestScore < 100")
+    Long countByUserId(@Param("userId") Long userId);
 }
