+++ src/main/java/com/potato/balbambalbam/user/join/service/JoinService.java
@@ -5,7 +5,7 @@ import com.potato.balbambalbam.data.entity.User;
 import com.potato.balbambalbam.data.entity.UserLevel;
 import com.potato.balbambalbam.data.repository.*;
 import com.potato.balbambalbam.exception.UserNotFoundException;
-import com.potato.balbambalbam.user.profile.dto.EditResponseDto;
+import com.potato.balbambalbam.user.setting.dto.EditResponseDto;
 import com.potato.balbambalbam.user.join.dto.JoinResponseDto;
 import com.potato.balbambalbam.user.token.jwt.JWTUtil;
 import jakarta.servlet.http.HttpServletResponse;
